feat: moved logic to detail page component and add routing (EC-97 EC-108)
